[Refactor] Terminal output drawer shared component
- Converts duplicate drawer terminal output code to a shared single file component
